Time: 0 ms (100.00%) | Memory: 8.5 MB (54.38%) - LeetSync

+++ 7-reverse-integer/reverse-integer.cpp
@@ -11,7 +11,7 @@ public:
         if (rev > INT_MAX || rev < INT_MIN)
             return 0;
 
-        return rev;
+        return (int) rev;
 
         
     }
